fix(calculate_ims_ascii): fix script and add tests
